Extract MultiSelect component
